requirements.txt
sqlmodel
httpx
fastapi[standard]
requests
feedparser
dotenv
apscheduler
newspaper3k
googlenewsdecoder
sentence-transformers
